Compile fixes picked up by VS6.  Including a SDB free() fix that was causing exceptions, but only when running in debugger. Don't know why never caught this one before.

+++ src/map.c
@@ -879,7 +879,7 @@ map (void)
                 lastRedraw[i] = 0;
             }
 
-            if (lastRedraw[i] == server_ups)
+            if (lastRedraw[i] == (u_int) server_ups)
             {
                 /* Redraw stationary ships every update so that these
                    ships are not hidden by planet updates. */
